third_party/blink/web_tests/http/tests/security/promise-access-control-allow-expected.txt

CONSOLE ERROR: line 1: Uncaught (in promise) 42
CONSOLE MESSAGE: line 22: PASS: received unhandledrejection event
CONSOLE MESSAGE: line 25: PASS: received expected reason
Tests that promise rejection events are received for cross origin CORS scripts
